ejercicio2/pyproject.toml
[project]
name = "food-service-api"
version = "0.1.0"
description = "API REST protegida con OAuth2"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.103.0",
    "uvicorn>=0.23.2",
    "sqlalchemy>=2.0.20",
    "pydantic>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "greenlet>=3.0.0",
    "httpx>=0.24.1",
    "aiosqlite>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "ruff>=0.9.4",
]

[tool.ruff]
line-length = 88
target-version = "py39"
respect-gitignore = true

[tool.ruff.lint]
extend-select = [
    "E",   # Pycodestyle
    "W",   # Pycodestyle Warnings
    "F",   # Pyflakes
    "B",   # flake8-bugbear
    "I",   # isort
    "N",   # pep8-naming
    "S",   # flake8-bandit
]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = true
